+++ Sources/ZcashLightClientKit/Error/ZcashErrorCode.swift
@@ -197,6 +197,8 @@ public enum ZcashErrorCode: String {
     case rustDeriveArbitraryAccountKey = "ZRUST0066"
     /// Error from rust layer when calling ZcashRustBackend.importAccountUfvk
     case rustImportAccountUfvk = "ZRUST0067"
+    /// Error from rust layer when calling ZcashRustBackend.deriveAddressFromUfvk
+    case rustDeriveAddressFromUfvk = "ZRUST0068"
     /// SQLite query failed when fetching all accounts from the database.
     case accountDAOGetAll = "ZADAO0001"
     /// Fetched accounts from SQLite but can't decode them.