+++ Sources/ZcashLightClientKit/Tool/DerivationTool.swift
@@ -17,6 +17,11 @@ public protocol KeyValidation {
 }
 
 public protocol KeyDeriving {
+    /// Derives and returns a UnifiedAddress from a UnifiedFullViewingKey
+    /// - Parameter ufvk: UTF-8 encoded String to validate
+    /// - Returns: true `UnifiedAddress`
+    func deriveUnifiedAddressFrom(ufvk: String) throws -> UnifiedAddress
+    
     /// Given the seed bytes and ZIP 32 account index, return the corresponding UnifiedSpendingKey.
     /// - Parameter seed: `[Uint8]` seed bytes
     /// - Parameter accountIndex: the ZIP 32 index of the account
@@ -92,6 +97,10 @@ public class DerivationTool: KeyDeriving {
         ZcashKeyDerivationBackend.getAddressMetadata(addr)
     }
 
+    public func deriveUnifiedAddressFrom(ufvk: String) throws -> UnifiedAddress {
+        try backend.deriveUnifiedAddressFrom(ufvk: ufvk)
+    }
+    
     /// Given a spending key, return the associated viewing key.
     /// - Parameter spendingKey: the `UnifiedSpendingKey` from which to derive the `UnifiedFullViewingKey` from.
     /// - Returns: the viewing key that corresponds to the spending key.